requirements.txt
django==5.0.3
python-dotenv==1.0.1
mysqlclient==2.2.4
Pillow==10.2.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
django-cors-headers ==4.3.1
django-filter==24.1